# services/go-escrow/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "State-machine escrow microservice with idempotency, rate limiting, and blockchain event auditting.",
        "title": "Go Escrow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Marketplace Support",
            "email": "marketplace@marketplace.local"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/v1/escrow",
    "paths": {
        "/v1/escrow": {
            "post": {
                "description": "Creates a new escrow account linked to an order with the specified amount.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Create a new escrow account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key (UUID)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Create escrow request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createEscrowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Escrow account created",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID or VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}": {
            "get": {
                "description": "Retrieves an escrow account by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Get escrow account by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Escrow account details",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}/advance": {
            "post": {
                "description": "Advances the escrow account to the next status (FUNDED → IN_PROGRESS → COMPLETED).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Advance escrow status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Advance request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.advanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Escrow status advanced",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID or VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}/cancel": {
            "post": {
                "description": "Cancels a funded or in-progress escrow account. Balance is zeroed and a cancel event is emitted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Cancel an escrow account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key (UUID)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Escrow account cancelled",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}/dispute": {
            "post": {
                "description": "Opens a dispute for a completed escrow account. Reason is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Dispute an escrow account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key (UUID)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Dispute request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.disputeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispute opened",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID or VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "VALIDATION_ERROR (reason required)",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}/fund": {
            "post": {
                "description": "Adds funds to an existing escrow account. Rate-limited per IP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Fund an escrow account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key (UUID)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Fund request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.fundEscrowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Escrow account funded",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID or VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "RATE_LIMITED",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/escrow/{id}/release": {
            "post": {
                "description": "Releases funds from the escrow account to the seller.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "escrow"
                ],
                "summary": "Release an escrow account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Escrow account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key (UUID)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Escrow account released",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_UUID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_UUID"
                },
                "detail": {
                    "type": "string",
                    "example": "id must be a valid UUID"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorItem"
                    }
                }
            }
        },
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "api.advanceRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "IN_PROGRESS",
                        "COMPLETED"
                    ],
                    "example": "IN_PROGRESS"
                }
            }
        },
        "api.createEscrowRequest": {
            "type": "object",
            "required": [
                "amount",
                "order_id"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "100.0000"
                },
                "order_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "api.disputeRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "Service not delivered as described"
                }
            }
        },
        "api.fundEscrowRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "100.0000"
                }
            }
        }
    },
    "securityDefinitions": {
        "X-Idempotency-Key": {
            "type": "apiKey",
            "name": "X-Idempotency-Key",
            "in": "header"
        },
        "X-Request-ID": {
            "type": "apiKey",
            "name": "X-Request-ID",
            "in": "header"
        }
    }
}